[REFACTOR] redux로 전역 상태 관리
