+++ src/Router.jsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import KaKaoMap from './components/location/KaKaoMap';
 import HomePage from './pages/HomePage';
 import InterviewPage from './pages/InterviewPage';
 import LocationPage from './pages/LocationPage';
@@ -17,7 +16,6 @@ function Router() {
           <Route path="/location" element={<LocationPage />} />
           <Route path="/interview" element={<InterviewPage />} />
         </Route>
-        <Route path="/test-location" element={<KaKaoMap />} />
       </Routes>
     </BrowserRouter>
   );
